+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 * Handling mimetypes other than `text/gemini`. Clients would have to download them instead of trying to display them. Example app is amended.
 * Added mimetype of response to the access log.
 * Added documentation for Handlers.
+* Redirect clients when pointing at a static subdirectory without a trailing slash. It caused misdirections because the client was requesting the "/document.gmi" instead of the "/subdir/document.gmi".
 
 ## v0.0.1 (2020-12-04)